Split circle poly module
