+++ src/core/fields/qm31.rs
@@ -75,7 +75,7 @@ mod tests {
     use rand::Rng;
 
     use super::QM31;
-    use crate::core::fields::m31::{M31, P};
+    use crate::core::fields::m31::P;
     use crate::core::fields::IntoSlice;
     use crate::m31;
